refactor(workspace): inline table schemas, split fuji actions
The tables/kv object literals in each app's workspace.ts had no external
consumers; each was a named constant that existed only to be passed back
into createXWorkspace() one line later. Inlining puts the schema in
exactly one place and removes the unused FujiTables, HoneycrispTables,
ZhongwenTables, and OpensidianTables type aliases.

Split Fuji's ~240-line action registry into apps/fuji/src/lib/actions.ts
for cohesion. The @epicenter/fuji package surface stays identical via a
re-export at the bottom of workspace.ts, so external consumers (examples,
docs, browser, daemon) need no changes.

Routed playground opensidian sites through createOpensidianWorkspace()
instead of re-implementing the bundle with raw opensidianTables; that
was the only external consumer keeping the export alive.
